refactor: move datatype stuff into gay.oss.gatos.core.data
