+++ core/src/test/java/gay/oss/gatos/core/graph/test/GraphTest.java
@@ -16,8 +16,8 @@ import gay.oss.gatos.core.graph.Node;
 import gay.oss.gatos.core.graph.NodeMetadata;
 import gay.oss.gatos.core.graph.connector.NodeConnection;
 import gay.oss.gatos.core.graph.connector.NodeConnector;
-import gay.oss.gatos.core.graph.data.DataBox;
-import gay.oss.gatos.core.graph.data.DataType;
+import gay.oss.gatos.core.data.DataBox;
+import gay.oss.gatos.core.data.DataType;
 import gay.oss.gatos.core.graph.type.NodeType;
 
 public class GraphTest {
