+++ core/src/test/java/gay/oss/gatos/core/graph/test/NodeTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.Test;
 
 import gay.oss.gatos.core.graph.Node;
 import gay.oss.gatos.core.graph.connector.NodeConnector;
-import gay.oss.gatos.core.graph.data.DataBox;
-import gay.oss.gatos.core.graph.data.DataType;
+import gay.oss.gatos.core.data.DataBox;
+import gay.oss.gatos.core.data.DataType;
 import gay.oss.gatos.core.graph.type.NodeType;
 
 public class NodeTest {
